Separated python code to packages
