+++ backend/bot/test/test_bot.py
@@ -1,4 +1,4 @@
-from bot import Bot
+from bot.bot import Bot
 
 
 def test_bot_create():
